Move xlator to bin/glusterd2/xlator
