+++ bin/glusterd2/commands/volumes/common.go
@@ -9,8 +9,8 @@ import (
 	"github.com/gluster/glusterd2/bin/glusterd2/transaction"
 	"github.com/gluster/glusterd2/bin/glusterd2/volgen"
 	"github.com/gluster/glusterd2/bin/glusterd2/volume"
+	"github.com/gluster/glusterd2/bin/glusterd2/xlator"
 	"github.com/gluster/glusterd2/pkg/utils"
-	"github.com/gluster/glusterd2/xlator"
 
 	"github.com/pborman/uuid"
 )
