+++ bin/glusterd2/main.go
@@ -12,10 +12,10 @@ import (
 	"github.com/gluster/glusterd2/bin/glusterd2/servers"
 	"github.com/gluster/glusterd2/bin/glusterd2/store"
 	"github.com/gluster/glusterd2/bin/glusterd2/volgen"
+	"github.com/gluster/glusterd2/bin/glusterd2/xlator"
 	"github.com/gluster/glusterd2/daemon"
 	"github.com/gluster/glusterd2/pkg/utils"
 	"github.com/gluster/glusterd2/version"
-	"github.com/gluster/glusterd2/xlator"
 
 	log "github.com/sirupsen/logrus"
 	flag "github.com/spf13/pflag"
