+++ bin/glusterd2/xlator/global.go
